test(smoke): compile-check example plugin fixture in `make test`
The embedded example-plugin source is built at runtime only by the
smoke-tagged plugin workflow tests, so SDK changes that break it slip past
`make test` and surface only in the slow golden-update workflow (twice
recently). Move examplePluginMainSource into a build-tag-free
fixture_compile_test.go and add TestExamplePluginFixtureCompiles, which builds
it against the live sdk via a throwaway replace module. It runs as part of
`go test ./...`, so fixture/SDK drift now fails fast. The smoke workflow reuses
the same const for its end-to-end runs.
